requirements.txt
streamlit
openai
python-dotenv
python-docx
requests
pandas
beautifulsoup4
python-dateutil
